+++ src/main/java/net/yxiao233/realmofdestiny/ModRegistry/ModBlockEntities.java
@@ -4,6 +4,7 @@ import net.minecraft.world.level.block.entity.BlockEntityType;
 import net.minecraftforge.registries.DeferredRegister;
 import net.minecraftforge.registries.ForgeRegistries;
 import net.minecraftforge.registries.RegistryObject;
+import net.yxiao233.realmofdestiny.Entities.CreativeEnergyMatrixBlockEntity;
 import net.yxiao233.realmofdestiny.Entities.GemPolishingStationBlockEntity;
 import net.yxiao233.realmofdestiny.Entities.PedestalBlockEntity;
 import net.yxiao233.realmofdestiny.RealmOfDestiny;
@@ -20,5 +21,7 @@ public class ModBlockEntities {
             BlockEntityType.Builder.of(PedestalBlockEntity::new,ModBlocks.PEDESTAL.get())
                     .build(null));
 
-
+    public static final RegistryObject<BlockEntityType<CreativeEnergyMatrixBlockEntity>> CREATIVE_ENERGY_MATRIX_BE = BLOCK_ENTITIES.register("creative_energy_matrix", () ->
+            BlockEntityType.Builder.of(CreativeEnergyMatrixBlockEntity::new,ModBlocks.CREATIVE_ENERGY_MATRIX.get())
+                    .build(null));
 }