+++ src/main/java/net/yxiao233/realmofdestiny/ModRegistry/ModBlocks.java
@@ -7,10 +7,7 @@ import net.minecraft.world.level.block.state.BlockBehaviour;
 import net.minecraftforge.registries.DeferredRegister;
 import net.minecraftforge.registries.ForgeRegistries;
 import net.minecraftforge.registries.RegistryObject;
-import net.yxiao233.realmofdestiny.Blocks.GemPolishingStationBlock;
-import net.yxiao233.realmofdestiny.Blocks.ModFlammableBlock;
-import net.yxiao233.realmofdestiny.Blocks.ModFlammableRotatedPillarBlock;
-import net.yxiao233.realmofdestiny.Blocks.PedestalBlock;
+import net.yxiao233.realmofdestiny.Blocks.*;
 import net.yxiao233.realmofdestiny.RealmOfDestiny;
 import net.yxiao233.realmofdestiny.worldgen.tree.BoltTreeGrower;
 
@@ -38,6 +35,8 @@ public class ModBlocks {
     public static final RegistryObject<Block> BOLT_PLANKS = BLOCKS.register("bolt_planks", () ->
             new ModFlammableBlock(BlockBehaviour.Properties.copy(Blocks.OAK_PLANKS)));
 
+    public static final RegistryObject<Block> CREATIVE_ENERGY_MATRIX = BLOCKS.register("creative_energy_matrix", () ->
+            new CreativeEnergyMatrixBlock(BlockBehaviour.Properties.copy(Blocks.IRON_BLOCK)));
 
     private static <T extends Block> RegistryObject<T> registryBlock(String name, Supplier<T> block){
         RegistryObject<T> toReturn = BLOCKS.register(name,block);