+++ src/main/java/net/yxiao233/realmofdestiny/ModRegistry/ModCreativeModeTab.java
@@ -20,6 +20,7 @@ public class ModCreativeModeTab {
                 output.accept(ModItems.BOLT_SAPLING_ITEM.get());
                 output.accept(ModItems.STRIPPED_BOLT_LOG_ITEM.get());
                 output.accept(ModItems.BOLT_PLANKS_ITEM.get());
+                output.accept(ModItems.CREATIVE_ENERGY_MATRIX_ITEM.get());
             })
             .title(Component.translatable("itemGroup.realmofdestiny"))
             .build()