+++ src/main/java/net/yxiao233/realmofdestiny/datagen/ModBlockStateProvider.java
@@ -31,6 +31,7 @@ public class ModBlockStateProvider extends BlockStateProvider {
         blockItem(ModBlocks.STRIPPED_BOLT_LOG);
         leavesBlock(ModBlocks.BOLT_LEAVES);
         blockWithItem(ModBlocks.BOLT_PLANKS);
+        blockWithItem(ModBlocks.CREATIVE_ENERGY_MATRIX);
     }
 
     private void leavesBlock(RegistryObject<Block> registryObject){