+++ src/main/java/net/yxiao233/realmofdestiny/datagen/ModBlockTagGenerator.java
@@ -20,11 +20,13 @@ public class ModBlockTagGenerator extends BlockTagsProvider {
     protected void addTags(HolderLookup.Provider provider) {
         this.tag(BlockTags.NEEDS_STONE_TOOL)
                 .add(ModBlocks.GEM_POLISHING_STATION.get())
-                .add(ModBlocks.PEDESTAL.get());
+                .add(ModBlocks.PEDESTAL.get())
+                .add(ModBlocks.CREATIVE_ENERGY_MATRIX.get());
 
         this.tag(BlockTags.MINEABLE_WITH_PICKAXE)
                 .add(ModBlocks.GEM_POLISHING_STATION.get())
-                .add(ModBlocks.PEDESTAL.get());
+                .add(ModBlocks.PEDESTAL.get())
+                .add(ModBlocks.CREATIVE_ENERGY_MATRIX.get());
 
         this.tag(BlockTags.MINEABLE_WITH_AXE)
                 .add(ModBlocks.BOLT_PLANKS.get())