+++ src/main/java/net/yxiao233/realmofdestiny/datagen/loot/ModBlockLootTables.java
@@ -23,6 +23,7 @@ public class ModBlockLootTables extends BlockLootSubProvider {
         this.dropSelf(ModBlocks.BOLT_SAPLING.get());
         this.dropSelf(ModBlocks.STRIPPED_BOLT_LOG.get());
         this.dropSelf(ModBlocks.BOLT_PLANKS.get());
+        this.dropSelf(ModBlocks.CREATIVE_ENERGY_MATRIX.get());
         this.add(ModBlocks.BOLT_LEAVES.get(), block ->
                 createLeavesDrops(block,ModBlocks.BOLT_SAPLING.get(),NORMAL_LEAVES_SAPLING_CHANCES));
     }